C02P20-9 Remove SIP 1.X events
The state database was originally designed for deliveries of an essence
and sidecar pair on the FTP, also known as the legacy ingest. This has
been partially expanded to facilitate SIP deliveries on the S3. Although
the delivery format and location are vastly different, that information
is stored in the same table resulting in sparse data. Also, not all the
SIP ingest information was handled and stored in the database resulting in
incomplete state.

Information about the delivery of a SIP on S3 and its ingest status will
be stored in a separate database table. Remove handling of said events,
so that only the legacy ingest remains.